+++ src/JsonApi/Document/AbstractSimpleResourceDocument.php
@@ -3,47 +3,12 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Yin\JsonApi\Document;
 
-use WoohooLabs\Yin\JsonApi\Schema\Data\DataInterface;
-use WoohooLabs\Yin\JsonApi\Schema\Data\SingleResourceData;
-use WoohooLabs\Yin\JsonApi\Transformer\Transformation;
+use WoohooLabs\Yin\JsonApi\Schema\Document\AbstractSimpleResourceDocument;
 
-abstract class AbstractSimpleResourceDocument extends AbstractSuccessfulDocument
-{
-    abstract protected function getResource(): array;
+class_alias(AbstractSimpleResourceDocument::class, 'WoohooLabs\Yin\JsonApi\Document\AbstractSimpleResourceDocument');
 
-    protected function createData(): DataInterface
-    {
-        return new SingleResourceData();
-    }
-
-    protected function fillData(Transformation $transformation): void
-    {
-        $transformation->data->addPrimaryResource($this->getResource());
-    }
-
-    protected function getRelationshipMember(
-        string $relationshipName,
-        Transformation $transformation,
-        array $additionalMeta = []
-    ): ?array {
-        $relationship = $this->getRelationshipFromResource($this->getResource(), $relationshipName);
-        if ($relationship !== null) {
-            $transformation->data->addPrimaryResource($relationship);
-        }
-
-        return $relationship;
-    }
-
-    private function getRelationshipFromResource(array $resource, string $relationshipName): ?array
-    {
-        if (empty($resource["relationships"][$relationshipName])) {
-            return null;
-        }
-
-        if (is_array($resource["relationships"][$relationshipName]) === false) {
-            return null;
-        }
-
-        return $resource["relationships"][$relationshipName];
-    }
-}
+trigger_error(
+    'Class WoohooLabs\Yin\JsonApi\Document\AbstractErrorDocument is deprecated, use ' .
+    AbstractSimpleResourceDocument::class . ' instead.',
+    E_USER_DEPRECATED
+);